chore(#659): Shape.swift stage 1, evict everything but extension Shape
Removes the 67 non-Shape standalone types, the 5 free functions that operate on
them (invisible to derive-swift-file-split.py's decl-only census, but still not
extension Shape), and the six foreign extension groups (Surface, Curve3D, Face,
Curve2D, Edge, Wire) from Shape.swift. The Shape class itself and every
extension Shape block stay, per the issue.

A one-shot Python migration script (parsed the file with the same DECL regex
and brace-depth scan as Scripts/derive-swift-file-split.py, attached each
declaration's leading /// doc-comment block, routed by a name->file map, wrote
the files) did the move, verified against the census, then was deleted.

One access-control fix was required, not optional: ShapeHistoryRef's init was
fileprivate, called only from extension Shape (which stays in Shape.swift) but
the class itself moves to BooleanResult.swift. fileprivate is file-scoped, so
the call sites would stop compiling. Widened to internal (removed the keyword,
matching the sibling handle property's own implicit-internal level) -- not a
public API change, since fileprivate and internal are both invisible outside
the module.
