+++ Sources/OCCTSwift/TangentialDeflectionPoint.swift
@@ -1,0 +1,9 @@
+import Foundation
+import simd
+import OCCTBridge
+
+/// A sampled point from tangential deflection.
+public struct TangentialDeflectionPoint {
+    public let parameter: Double
+    public let x: Double, y: Double, z: Double
+}
